+++ src/main/java/edu/metrostate/ics499/team2/model/Flashcard.java
@@ -1,6 +1,5 @@
 package edu.metrostate.ics499.team2.model;
 
-import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document(collection = "flashcard")
